% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Weighted leveraging Regression}
\usage{
algo_leverage(y, x, r, method)
}
\arguments{
\item{y}{response variable}

\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector}

\item{r}{number of subsamples}

\item{method}{"weighted" or "unif"}
}
\value{
a fitted linear regression
}
\description{
Weighted leveraging Regression
}
\examples{
n=500
x<-rt(500,df=6)
noise<-rnorm(500)
y<--x+noise
algo_leverage(y,x,r=100,method="weighted")

}
