% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasso.R
\name{elasso}
\alias{elasso}
\title{Title}
\usage{
elasso(X, y, nlambda = 100, alpha, thresh = 1e-05, maxiter = 10000)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation vector}

\item{y}{response variable}

\item{nlambda}{The number oflambdavalues - default is 100}

\item{alpha}{The elasticnet mixing paramete}

\item{thresh}{Convergence threshold for coordinate descent.}

\item{maxiter}{}
}
\value{
length(lambda)matrix of coefficients
}
\description{
Title
}
