% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{HWGen}
\alias{HWGen}
\title{Title
Generating 3 banded matrix to solve the hw1}
\usage{
HWGen(n, alpha)
}
\arguments{
\item{n}{dimension of the square matrix}

\item{alpha}{the diagonal value}
}
\value{
3 banded matrix
}
\description{
Title
Generating 3 banded matrix to solve the hw1
}
