% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{plot_el}
\alias{plot_el}
\title{plot the elnet_coord}
\usage{
plot_el(fit)
}
\arguments{
\item{fit}{the fitted object via elnet_coord}
}
\value{
a plot
}
\description{
plot the elnet_coord
}
\examples{

}
