% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Title}
\usage{
solve_ols(A, v, method, ncores = 1, iter.max = 10000)
}
\arguments{
\item{A}{the linear matrix of coefficient}

\item{v}{the underlying true solution}

\item{method}{"G-S" stands for Gauss-Seidel method and "Jacobi" uses Jacobi method}

\item{ncores}{1 stands for non-parallel computing and usually ncores<=4}

\item{iter.max}{the maximal number of iteration}
}
\value{
relative error: the relative error at each iteration
        solution
}
\description{
Apply Gauss-Seidel or Jacobi method to solve the linear equations
}
\examples{
n=100
alpha=3
A<-HWGen(n,alpha)
v<-rep(c(1,0),50)
solve_ols(A,v,method="G-S")

}
