% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasso.R
\name{elasso}
\alias{elasso}
\title{Title}
\usage{
elasso(X, y, nlambda = 100, alpha, thresh = 1e-05, maxiter = 10000)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation vector}

\item{y}{response variable}

\item{nlambda}{The number oflambdavalues - default is 100}

\item{alpha}{The elasticnet mixing paramete}

\item{thresh}{Convergence threshold for coordinate descent.}

\item{maxiter}{the maximal number of iterations in each coordinate descent}
}
\value{
length(lambda)matrix of coefficients
}
\description{
Solve the an elastic net problem via coordinate descent
}
\examples{
n=50
p=20
beta<-c(2,0,-2,0,1,0,-1,0,rep(0,12))
X=matrix(rnorm(n*p),n,p)
Y<-X\%*\%beta+rnorm(n)
fit1<-elasso(X,Y,alpha=0.3)
beta_cd<-t(as.matrix(fit1$beta))
pct <- rowSums(abs(beta_cd))
matplot(pct,beta_cd,type="l",lty=1,
       xlab="|beta|_1",ylab="Coefficients")
text(max(pct)+0.2,beta_cd[dim(beta_cd)[1],],1:p,col=1:p,cex=0.7)


}
